src/lib/server/requirements.txt
fastapi==0.112.2 
uvicorn==0.34.0
python-multipart==0.0.9
numpy==1.26.3
pillow==10.2.0
torch==2.5.1
torchvision==0.20.1